[FEATURE] Add 'form styling' select field in form editor
This is an example how to extend the form editor in the TYPO3 backend
with a custom field.

The select field contains three options for different styling classes.
The selected class is rendered in the customized Fluid template of
the Form.
This example does NOT provide a stylesheet. Therefore, the classes
have no visible effect in the frontend.