[FEATURE] Replace custom HTML mail example with FluidEmail
When the TYPO3 Form Framework was introduced with TYPO3 v8,
the HTML emails sent only contained a table of form contents.

It was the integrator's task to implement a custom template
with a logo, footer etc. for more appealing emails.

In version 10.3, the TYPO3 core introduced Fluid-based email templating.
Support for EXT:form was added with 10.4.

This allows us to replace the former custom email template with the
'useFluidEmail' option in the email finisher.

The updated example shows how to implement a custom Fluid template
for all emails with this option enabled, as well as a special
Fluid template (and layout) for selected email finishers.