[TASK] Set root paths for FluidEmail in Form configuration
The former registration of template paths via
$GLOBALS['TYPO3_CONF_VARS']['MAIL'] worked globally. This caused other
emails sent by TYPO3 to use the customized 'SystemEmail' layout.
While this is not necessarily bad, the examples in this extension
should only affect the TYPO3 Form Framework.

Using the YAML configuration, the custom root paths are now limited to
the Email finishers of EXT:form.